project_root/requirements.txt
flask==2.3.3
flask-wtf==1.1.1
flask-login==0.6.2
python-docx==0.8.11
PyPDF2==3.0.1
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7